fix iterator
Sadly iterator API needed to be changed slightly.

Instead of operator->() returning a pointer to pair<key_type,mapped_type> it
needs to return a unique_ptr to a pair<const key_type,mapped_type&>.
It behaves the same, but is technically different.

Instead of operator*() returning a reference to pair<key_type,mapped_type>
it returns a pair<const key_type,mapped_type&> with only a reference to
the mapped object, but not to the key object any more. Again this will
behave the same but is technically different. This is so the hash can
be stored in place of the key when there is an unhash that defines the
bijection.